Move Android string resources code into new module
